+++ lib/src/model/ics_status.dart
@@ -1,2 +1,12 @@
 /// Enumeration of all event statuses.
-enum IcsStatus { TENTATIVE, CONFIRMED, CANCELLED }
+/// See doc: https://www.kanzaki.com/docs/ical/status.html
+enum IcsStatus {
+  TENTATIVE,
+  CONFIRMED,
+  CANCELLED,
+  NEEDS_ACTION,
+  COMPLETED,
+  IN_PROCESS,
+  DRAFT,
+  FINAL
+}